requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
greenlet==3.2.0
idna==3.10
playwright==1.51.0
pyee==12.1.1
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
typing_extensions==4.13.2
urllib3==2.4.0
